requirements.txt
# Core LLM & RAG
langchain==0.1.20
langchain-groq==0.1.3
langchain-community==0.0.38

# Vector Database
chromadb==0.4.22

# PDF Processing
pypdf==4.0.1

# Web Framework
fastapi==0.109.0
uvicorn==0.27.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3

# Data Processing
numpy==1.26.3
